fix: move not exported utils to util file
