+++ packages/playback-core/src/util.ts
@@ -1,3 +1,5 @@
+import { ValueOf, StreamTypes, ExtensionMimeTypeMap, MimeTypeShorthandMap, MuxMediaProps } from './types';
+
 type KeyTypes = string | number | symbol;
 
 type addEventListenerWithTeardown = <K extends keyof HTMLMediaElementEventMap>(
@@ -38,3 +40,64 @@ export function inSeekableRange(seekable: TimeRanges, duration: number, time: nu
   }
   return false;
 }
+
+export const toPlaybackIdParts = (playbackIdWithOptionalParams: string): [string, string?] => {
+  const qIndex = playbackIdWithOptionalParams.indexOf('?');
+  if (qIndex < 0) return [playbackIdWithOptionalParams];
+  const idPart = playbackIdWithOptionalParams.slice(0, qIndex);
+  const queryPart = playbackIdWithOptionalParams.slice(qIndex);
+  return [idPart, queryPart];
+};
+
+export const getType = (props: Partial<Pick<MuxMediaProps, 'type' | 'src'>>) => {
+  const type = props.type;
+  if (type) {
+    const upperType = type.toUpperCase();
+    return isKeyOf(upperType, MimeTypeShorthandMap) ? MimeTypeShorthandMap[upperType] : type;
+  }
+
+  const { src } = props;
+  if (!src) return '';
+
+  return inferMimeTypeFromURL(src);
+};
+
+export const inferMimeTypeFromURL = (url: string) => {
+  let pathname = '';
+  try {
+    pathname = new URL(url).pathname;
+  } catch (e) {
+    console.error('invalid url');
+  }
+
+  const extDelimIdx = pathname.lastIndexOf('.');
+  if (extDelimIdx < 0) return '';
+
+  const ext = pathname.slice(extDelimIdx + 1);
+  const upperExt = ext.toUpperCase();
+
+  return isKeyOf(upperExt, ExtensionMimeTypeMap) ? ExtensionMimeTypeMap[upperExt] : '';
+};
+
+export const getStreamTypeConfig = (streamType?: ValueOf<StreamTypes>) => {
+  // for regular live videos, set backBufferLength to 8
+  if ([StreamTypes.LIVE, StreamTypes.DVR].includes(streamType as any)) {
+    const liveConfig = {
+      backBufferLength: 8,
+    };
+
+    return liveConfig;
+  }
+
+  // for LL Live videos, set backBufferLenght to 4 and maxFragLookUpTolerance to 0.001
+  if ([StreamTypes.LL_LIVE, StreamTypes.LL_DVR].includes(streamType as any)) {
+    const liveConfig = {
+      backBufferLength: 4,
+      maxFragLookUpTolerance: 0.001,
+    };
+
+    return liveConfig;
+  }
+
+  return {};
+};
